+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - improved logging
 - Added logging file with log rotation.
+- added default headers to /api endpoints => `Content-Type: application/json` `Accept: application/json`
 
 # Version 0.1.0 (26.02.2024)